pyproject.toml
[project]
name = "cleanshot"
version = "0.0.1"
description = "A tool to automatically rename screenshots using OpenAI's OCR"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Shlomo Dahan" }
]
dependencies = [
    "watchdog>=6.0.0",
    "openai>=1.76.0",
    "python-dotenv>=1.1.0",
    "Pillow>=11.2.1",
    "questionary>=1.10.0",
    "rumps>=0.4.0",
]

[project.scripts]
cleanshot = "cleanshot.main:app"

[project.optional-dependencies]
dev = [
    "ruff>=0.11.7"
]

[build-system]
requires = ["setuptools>=77.0.3"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["cleanshot*"]

[tool.ruff]
line-length = 120 

 